pyproject.toml
[tool.poetry]
name = "pyproject-toml"
version = "0.1.0"
description = ""
authors = ["Ilyas <92800413+ilyasnawawie@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "pyproject"}]


[tool.poetry.dependencies]
python = "^3.11"
django-environ = "^0.10.0"
celery = "^5.2.7"
pytest = "^7.3.1"
sentry-sdk = "^1.22.2"
black = "^23.3.0"
django-storages = "^1.13.2"
mysqlclient = "^2.1.1"
redis = "^4.5.5"
mysql-connector-python = "^8.0.26"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
